rehome preset-selector
